UBUNTU: rename the debug packages to match archive standard
BugLink: http://bugs.launchpad.net/bugs/527837

Our .ddebs are expiring after 14 days as the archive cleaner expects
the debug packages to have a matching binary package.  The prefix of the
debug package must match the full name of the binary package.  As we are
renaming them move to the archive standard for the postfix.  Rename the
debug packages as below:

    linux-image-VERSION-ABI-FLAVOUR-dbgsym

Signed-off-by: Andy Whitcroft <apw@canonical.com>