UBUNTU: update to standards version 3.8.4.0
Update from standards versions 3.6.1 to 3.8.4.0.  The following items
seem to be the only ones which could apply to the kernel in its current
form, comments inline:

  3.8.2.0
  --------

  Released Jun 2009.

  + *2.4*: The list of archive sections has been significantly expanded.  See
    [http://lists.debian.org/debian-devel-announce/2009/03/msg00010.html]
    for the list of new sections and rules for how to categorize
    packages.

Confirmed that the sections correctly match the archive overrides.

  3.8.1.0
  --------

  Released  Mar 2009.

  + *4.9.1*: New nocheck option for DEB\_BUILD\_OPTIONS indicating any build-time
    test suite provided by the package should not be run.

We have no test-suite for the built kernel, so this does not actually
apply.  The only tests we have are configuration tests.

  3.7.3.0
  --------

  Released  Dec 2007.

  + *12.5*: GFDL 1.2, GPL 3, and LGPL 3 are now in common-licenses and should
    be referenced rather than quoted in `debian/copyright'.

The copyright has already been updated to reference the common GPL-2
version.

Signed-off-by: Andy Whitcroft <apw@canonical.com>