ARM: omap5: sata: sata hwmod update
The sys config base address of sata conntroller is added to
hwmod of the sata.
The PLL base adress and sysconfig address of sata phy is added to
hwmod of the sata. this will be the temproray inclusion
once the seperate sata phy driver is aviable the pll and
sysconfig address should be removed.

Signed-off-by: Keshava Munegowda <keshava_mgowda@ti.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>