OMAP: DVFS: Adding volt request for dep domain should be based on user value
The volt request for dependent domain should be added based on the user requested value but _not_ from the plist of the voltdm.

There is a corner case issue, where two req_dev which share the same voltdm does device_scale.
And that voltdm has dependency on other voltdm. In this scenario,
one of the req_dev will end up not releasing the request from the _dependent_ volt domain.

For example:
rpres requests for 266MHz => 1.14V on IVA.
aess requests  for 196MHz => 1.14V on IVA.

If we follow the code flow. rpres will add user request on IVA voltdm. As IVA has
dependency on CORE, this will add a user request on CORE. And something similar for aess
as well. aess adds user requst on IVA and IVA adds user request on CORE.

So far so good. Now, rpress releases the request by requesting OPP50 freq. Assume, aess didn't
release the request yet. So the code flow will be, rpress requests for OPP50 freq and updates
the existing user request to OPP50 volt on IVA. However, IVA ends up updating user request for OPP100
on CORE instead of OPP50. Because the dependent domain volt request is taken from the plist. And in
the plist for IVA voltdm the higest volt is still OPP100, as aess is still active. So, this will
end up rpress to have OPP100 request on CORE voltdm always.

Fix:
Move the scaning of dependent volt at the early stage in device_scale and update the request
for dependent domain with user requested rate and not the plist value.

Signed-off-by: Girish S G <girishsg@ti.com>
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>