omap3+: abb: Adaptive Body-Bias structures & data
Starting with OMAP3630, some voltage domains have an LDO for controlling
voltage bias, called the Adaptive Body-Bias LDO or ABB.  The purpose of
this LDO is two-fold: boost voltage to PMOS backgates at high OPP for
weak silicon, and reduce voltage at lower OPPs on strong silicon.

The former use mentioned above is the Forward Body-Bias mode, or FBB.
The latter use is Reverse Body-Bias or RBB.  ABB can also be bypassed
(the default state) in which case it will just follow the voltage of the
VP/VC.

This patch introduces the data structures needed to represent the ABB
ldo's in the voltage layer, and populates the appropriate data for 3630
and OMAP4.  Not all voltage domains have an ABB ldo, and OMAP34xx does
not have it at all; in such cases the voltage data will be marked with
OMAP_ABB_NO_LDO.

Signed-off-by: Mike Turquette <mturquette@ti.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>