HACK: usb: mass_storage: Increase write throughput by 2x on musb
Currently the musb controller driver relies on short_not_ok flag in the
usb request to use DMA mode 1 which gives a performance boost.
We set this flag in the mass storage gadget to take benefit of DMA mode
1. This would give us 2x performance boost in PC to device transfers in
mass storage mode.

This patch is a HACK because it is not a generic solution and would
break USB3.0 controllers using mass storage mode. We intend to revert it
and use a proper solution when it is ready.

Signed-off-by: Roger Quadros <rogerq@ti.com>