UBUNTU: Add compat-wireless-3.6
BugLink: http://bugs.launchpad.net/bugs/1067543

Replace obsolete cw-3.4 with cw-3.6.  Bug 1067543 also specifically
requested the alx driver which is not officially upstream.  The
gen-stable-release.sh script in the compat-drivers tree does provide an
option (-c flag) when creating the compat-wireless snapshots.  Thus, we
speicifcally pulled the tarball which used this -c flag in order to
provide this driver:

  http://www.orbit-lab.org/kernel/compat-wireless-3-stable/v3.6/compat-wireless-3.6.2-1-snpc.tar.bz2

For more information about generating the snapshot:

  http://linuxwireless.org/en/users/Download/hacking#Generating_stable_releases

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>