UBUNTU: Update linux-crashdump dependencies
We were using some additional scripts in kexec-tools to produce
kernel dumps via kexec. Though upstream and Debian got tools
to create those dumps packaged as kdump-tools. We want to reduce
the deviation from Debian, so changing the meta package to depend
on kdump-tools.

The kdump-tools package depends on kexec-tools and makedumpfile,
so those two can be dropped from the dependencies.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>